+++ lib/secretary.rb
@@ -31,6 +31,7 @@ module Secretary
   autoload :HasSecretary
   autoload :VersionedAttributes
   autoload :TracksAssociation
+  autoload :Dirty
 end
 
 ActiveSupport.on_load(:active_record) do