+++ lib/secretary/dirty.rb
@@ -1,0 +1,8 @@
+module Secretary
+  module Dirty
+    extend ActiveSupport::Autoload
+
+    autoload :CollectionAssociation
+    autoload :SingularAssociation
+  end
+end