+++ spec/factories.rb
@@ -5,6 +5,17 @@ FactoryGirl.define do
     color "gray"
   end
 
+  factory :car do
+    name "Betsy"
+    color "white"
+    year 1984
+  end
+
+  factory :image do
+    title "Obama"
+    url "http://obama.com/obama.jpg"
+  end
+
   factory :location do
     title "Crawford Family Forum"
     address "474 S. Raymond, Pasadena"