+++ spec/internal/app/models/image.rb
@@ -1,0 +1,6 @@
+class Image < ActiveRecord::Base
+  has_secretary on: ["title", "url"]
+
+  belongs_to :story
+  tracks_association :story
+end