+++ spec/internal/app/models/story.rb
@@ -1,3 +1,7 @@
 class Story < ActiveRecord::Base
   has_secretary
+
+  has_one :image
+  accepts_nested_attributes_for :image, allow_destroy: true
+  tracks_association :image
 end