+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 * Added better way to declare attribute inclusion/exclusion:
   `has_secretary on: []` and `has_secretary except: []`. This ensures that
   attributes will be declared before any `tracks_association` are called.
+* Added support for singular associations tracking (has_one, belongs_to)
+
+
 ## 1.0.0.beta1 (2013-11-05)
 * Beta for stable 1.0 release.